requirements.txt
django==5.1.2
django-cors-headers==4.4.0
django-environ==0.11.2
django-seeding==1.3.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
numpy==2.1.2
pandas==2.2.3
pillow==10.4.0
psycopg2-binary==2.9.9
django_csp==4.0b1
openai==1.52.0
gtts==2.5.3
dj-database-url==2.3.0
'whitenoise[brotli]'
